+++ apps/backend/src/app/api/v1/users/crud.tsx
@@ -1,4 +1,4 @@
-import { ensureTeamMembershipExists, ensureUserExist } from "@/lib/request-checks";
+import { ensureTeamMembershipExists, ensureUserExists } from "@/lib/request-checks";
 import { PrismaTransaction } from "@/lib/types";
 import { sendTeamMembershipDeletedWebhook, sendUserCreatedWebhook, sendUserDeletedWebhook, sendUserUpdatedWebhook } from "@/lib/webhooks";
 import { prismaClient } from "@/prisma-client";
@@ -39,22 +39,6 @@ export const userFullInclude = {
     },
   },
 } satisfies Prisma.ProjectUserInclude;
-
-export const contactChannelToCrud = (channel: Prisma.ContactChannelGetPayload<{}>) => {
-  // eslint-disable-next-line @typescript-eslint/no-unnecessary-condition
-  if (channel.type !== 'EMAIL') {
-    throw new StackAssertionError("Only email channels are supported");
-  }
-
-  return {
-    id: channel.id,
-    type: 'email',
-    value: channel.value,
-    is_primary: !!channel.isPrimary,
-    is_verified: channel.isVerified,
-    used_for_auth: !!channel.usedForAuth,
-  };
-};
 
 export const oauthProviderConfigToCrud = (
   config: Prisma.OAuthProviderConfigGetPayload<{ include: {
@@ -481,7 +465,7 @@ export const usersCrudHandlers = createLazyProxy(() => createCrudHandlers(usersC
   },
   onUpdate: async ({ auth, data, params }) => {
     const result = await prismaClient.$transaction(async (tx) => {
-      await ensureUserExist(tx, { projectId: auth.project.id, userId: params.user_id });
+      await ensureUserExists(tx, { projectId: auth.project.id, userId: params.user_id });
 
       if (data.selected_team_id !== undefined) {
         if (data.selected_team_id !== null) {
@@ -768,7 +752,7 @@ export const usersCrudHandlers = createLazyProxy(() => createCrudHandlers(usersC
   },
   onDelete: async ({ auth, params }) => {
     const { teams } = await prismaClient.$transaction(async (tx) => {
-      await ensureUserExist(tx, { projectId: auth.project.id, userId: params.user_id });
+      await ensureUserExists(tx, { projectId: auth.project.id, userId: params.user_id });
 
       const teams = await tx.team.findMany({
         where: {