+++ packages/stack-sc/CHANGELOG.md
@@ -1,5 +1,27 @@
 # @stackframe/stack-sc
 
+## 2.6.8
+
+## 2.6.7
+
+### Patch Changes
+
+- Bugfixes
+
+## 2.6.6
+
+## 2.6.5
+
+### Patch Changes
+
+- Minor improvements
+
+## 2.6.4
+
+### Patch Changes
+
+- fixed small problems
+
 ## 2.6.3
 
 ## 2.6.2