+++ packages/stack-shared/CHANGELOG.md
@@ -1,5 +1,41 @@
 # @stackframe/stack-shared
 
+## 2.6.8
+
+### Patch Changes
+
+- Bugfixes
+  - @stackframe/stack-sc@2.6.8
+
+## 2.6.7
+
+### Patch Changes
+
+- Updated dependencies
+  - @stackframe/stack-sc@2.6.7
+
+## 2.6.6
+
+### Patch Changes
+
+- @stackframe/stack-sc@2.6.6
+
+## 2.6.5
+
+### Patch Changes
+
+- Minor improvements
+- Updated dependencies
+  - @stackframe/stack-sc@2.6.5
+
+## 2.6.4
+
+### Patch Changes
+
+- fixed small problems
+- Updated dependencies
+  - @stackframe/stack-sc@2.6.4
+
 ## 2.6.3
 
 ### Patch Changes